Be more resilient when reading pidfiles
Unfortunately, unicorn doesn't write its pidfile atomically, which means there
is a small but finite chance that during hot-reload, Unicorn Herder will read
the pidfile after the new master has opened it, but before it has
written/closed it. The result of this is that Unicorn Herder becomes detached
from the new master.

This commit changes that, by making Unicorn Herder attempt to read a pidfile
multiple times before giving up, provided it is not expecting the unicorn
process to exit.