Support .2 style pidfiles
When a new gunicorn process spawns, a new pidfile with a suffix of .2 is
created to contain the new PID. We need to be able to read from that
file so we can read the new PID of the master process and correctly shut
down the old process.

Note that this is different to unicorn which instead writes the PID to
the existing file and writes a new file with the suffix of .oldbin which
will contain the old PID.

Documentation for gunicorn: https://gunicorn-docs.readthedocs.io/en/latest/signals.html

> First, replace the old binary with a new one, then send a USR2 signal
> to the current master process. It executes a new binary whose PID file
> is postfixed with .2 (e.g. /var/run/gunicorn.pid.2), which in turn
> starts a new master process and new worker processes.

Documentation for unicorn: https://bogomips.org/unicorn/SIGNALS.html

> If you're using a pid file, the old process will have ".oldbin"
appended to its path.