Update Python versions
The world of Python has moved on a fair bit since this was last updated.
I've updated it to build against pypy3 and Python 3.8 as these are
current. I've left the check for 2.7 as I expect we (GDS) still run
Python 2.7.

In order for this to work with Python 3.8 the strings sent to file.write
needed to be converted to bytes, hence the encode change.